5.2.3 complete sorting and refactoring - добавил описание к файлам - завершил реализацию сортировки - мелкий рефакторинг кода
