--- src/const.js
@@ -1,20 +1,0 @@
-const COUNTRIES = [
-  'USA',
-  'ALBANIA',
-  'TAJIKISTAN',
-  'FINLAND',
-  'USSR',
-];
-
-const GENRES = [
-  'Musical',
-  'Drama',
-  'Thriller',
-  'Comedy',
-  'Historical',
-];
-
-export {
-  COUNTRIES,
-  GENRES
-};
