+++ src/mock/film.js
@@ -2,7 +2,7 @@ import dayjs from 'dayjs';
 import { nanoid } from 'nanoid';
 import { getRandomNumber, getRandomItem, shuffleArr } from '../utils/random.js';
 import { getRandomDate } from '../utils/days.js';
-import { COUNTRIES, GENRES } from '../const.js';
+import { COUNTRIES, GENRES } from '../utils/const.js';
 import { getComments } from './comments.js';
 
 const generateRandomBoolean = () => !!getRandomNumber(0, 1);
