+++ src/utils/common.js
@@ -1,0 +1,11 @@
+/*
+* разные методы для переиспользования в проекте
+* */
+
+const clearActiveClass = (element, className) => {
+  element.forEach((item) => item.classList.remove(`${ className }`));
+};
+
+export {
+  clearActiveClass
+};
