+++ src/utils/const.js
@@ -1,9 +1,31 @@
-const TypeSort = {
+/*
+* Неизменяемые значения в приложении
+* */
+
+const SortType = {
   DEFAULT: 'default',
   DATE: 'date',
   RATING: 'rating',
 };
 
+const COUNTRIES = [
+  'USA',
+  'ALBANIA',
+  'TAJIKISTAN',
+  'FINLAND',
+  'USSR',
+];
+
+const GENRES = [
+  'Musical',
+  'Drama',
+  'Thriller',
+  'Comedy',
+  'Historical',
+];
+
 export {
-  TypeSort
+  SortType,
+  COUNTRIES,
+  GENRES
 };
