+++ src/utils/days.js
@@ -1,3 +1,7 @@
+/*
+* библиотека для удобной работы с датами
+* */
+
 import dayjs from 'dayjs';
 import isToday  from 'dayjs/plugin/isToday';
 import isYesterday   from 'dayjs/plugin/isYesterday';
