+++ src/utils/observer.js
@@ -1,3 +1,7 @@
+/*
+* абстрактный паттерн наблюдатель
+* */
+
 const observer = {
   // ключ в котором хранятся observers (т.е слушатели)
   _observers: [],
