+++ src/view/sort.js
@@ -1,11 +1,11 @@
-/* eslint-disable */
 import Abstract from './abstract.js';
+import { SortType } from '../utils/const.js';
 
 const createSortingTemplate = () => (
   `<ul class="sort">
-    <li><a href="#" data-sort-type="default" class="sort__button sort__button--active">Sort by default</a></li>
-    <li><a href="#" data-sort-type="date" class="sort__button">Sort by date</a></li>
-    <li><a href="#" data-sort-type="rating" class="sort__button">Sort by rating</a></li>
+    <li><a href="#" data-sort-type="${ SortType.DEFAULT }" class="sort__button sort__button--active">Sort by default</a></li>
+    <li><a href="#" data-sort-type="${ SortType.DATE }" class="sort__button">Sort by date</a></li>
+    <li><a href="#" data-sort-type="${ SortType.RATING }" class="sort__button">Sort by rating</a></li>
   </ul>`
 );
 
@@ -21,13 +21,13 @@ export default class Sort extends Abstract {
   }
 
   _sortTypeChangeHandler(e) {
-    if (e.target !== e.target.closest('A')) return;
-    this._callback.sotrCallback(e.target)
+    if (e.target !== e.target.closest('A')) { return; }
+    this._callback.sotrCallback(e.target);
   }
 
   setSortTypeChangeHandler(callback) {
     this._callback.sotrCallback = callback;
-    this.getElement().addEventListener('click', this._sortTypeChangeHandler)
+    this.getElement().addEventListener('click', this._sortTypeChangeHandler);
   }
 }
